Using BigDecimals to calculate birthday problem.

+++ lib/src/test/java/com/aharpour/statistics/probability/BirthdayProblemTest.java
@@ -2,14 +2,16 @@ package com.aharpour.statistics.probability;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class BirthdayProblemTest {
 
     @Test
     void classicCase() {
-        double result = new BirthdayProblem(365, 23).getResult();
-        assertEquals(0.507297, result, 0.000001);
+        BigDecimal result = new BirthdayProblem(365, 23).getResult();
+        assertEquals(0.507297, result.doubleValue(), 0.000001);
     }
 
 }
